Update JS code to correctly preview PDFs
It appears that the original URLs used for importing pdf.js have
disappeared. This commit revises functionality by using a new URL, now
being built with JS modules in mind. Due to this, some of the code was
re-organized to work with the module system. This code is not (and never
was) a master class in best practices, and restructuring the JS to use
newer guidelines is beyond the scope of this page at this current time.
This is simply a patchwork fix to continue proper site behavior.